requirements.txt
Django==5.0.3
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
mysqlclient==2.2.4
django-environ
django-extensions